requirements.txt
jaconv==0.2.4
mecab-python3==0.996.2

